Feature: Last Attempted Url
* Permit function now records the lastAttemptedUrl that the user was
trying to navigate to before they were redirected due to
authentication/authorization issues. This value is stored in the storage
service so it can persist across any kind of redirects required for
authentication (in-memory storage won't work for this). When the user
logs in the last attempted URL is preferentially used over the default
onLoginRedirectPath if the trackLastAttemptedUrl setting is enabled.
* Corrected all uses of Path to Url and changed from $location.path() to
$location.url() usage, which allows for parameters and anchors to be
used/captured.