Add doNotRedirect parameter to logoutConfirmed method
logoutConfirmed will skip logout redirection if the doNotRedirect parameter is set to true. This helps prevent a bug in logout redirection that occurs when the user's profile is expired or the auth cookie is missing. The user is known, they just need fresh credentials and logoutConfirmed should not be redirecting the user as they are currently in a redirect flow to be authenticated.